requirements.txt
-e .
aws_cdk.aws_ecs==1.101.0
aws_cdk.aws_ec2==1.101.0